#108 concept creator to an extra helper object in test scope
